backend/app/core/config.py
import os
from dotenv import load_dotenv

load_dotenv()


class Settings:
    GROQ_API_KEY: str = os.getenv("GROQ_API_KEY")
    SUPABASE_DATABASE_URL: str = os.getenv("SUPABASE_DATABASE_URL")


settings = Settings()