backend/requirements.txt
fastapi
uvicorn
python-dotenv
groq
pydantic
sqlalchemy
asyncpg
ddgs